requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.66
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.10.0
fastapi>=0.110.0
uvicorn>=0.25.0 